+++ CHANGELOG.md
@@ -3,5 +3,17 @@
 Note the first digit of every adapter version corresponds to the major version of the Chartboost Mediation SDK compatible with that adapter. 
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
+### 4.10.1.3.1
+- Updated the dependency on Chartboost Mediation SDK to 4.0.0.
+
+### 4.10.1.3.0
+- This version of the adapter has been certified with InMobi SDK 10.1.3.
+
+### 4.10.1.1.2
+- Updated the dependency on Chartboost Mediation SDK to 4.0.0.
+
+### 4.10.1.1.1
+- Add impression callbacks.
+
 ### 4.10.1.1.0
 - This version of the adapter has been certified with InMobi SDK 10.1.1.